% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charToNum.R
\name{charToNum}
\alias{charToNum}
\title{A function to convert characters to numerics.}
\usage{
charToNum(charred_num)
}
\arguments{
\item{Any}{number with comma, dollar signs, pecent signs or spaces in it.}
}
\description{
This function converts money and other excel numeric variables that get coded as characters to numerics.
}
\examples{
charToNum("$1,204,444")
}
\keyword{numerics}
