% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColNames.R
\name{cleanColNames}
\alias{cleanColNames}
\title{Standardizes column names}
\usage{
cleanColNames(dat)
}
\description{
Sets column names to all lower case with underscores instead of spaces or periods.
}
\examples{
cleanColNames(head(iris))
}
