% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbexists.R
\name{dbexists}
\alias{dbexists}
\title{Table exists}
\usage{
dbexists(chan, tname)
}
\arguments{
\item{chan}{open RODBC channel}

\item{tname}{tablename (including schema if there is one)}
}
\description{
Checks for the existence of a database table
}
\examples{
dbtest <- odbcConnect('test')
sqlSave(dbtest, mtcars)
dbexists(dbtest, mtcars)
}
\keyword{tables}
